USB: android gadget: mass_storage: read/write performance enhancement
Implementing the csw hack feature & increasing the number of
buffers used in mass storage to 4.

As per the mass storage protocol, csw is sent from the device
after successfully writing the data on to storage media, where
there is a time factor involved. With this csw hack, csw is sent
just after receiving the data & before writing the data on the
storage media. By, any chance, if the write fails, the corresponding
lun is removed to indicate the user of write failure.

Buffers has been increased to 4, so that there is enough space for
the incoming data, which is usually 64KB per CBW.

User has the option of selecting/deselecting this feature from
the kernel config.